requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==2.0.0.post0
idna==3.4
magic-filter==1.0.9
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.0
SQLAlchemy==1.4.43
tomli==2.0.1
typing_extensions==4.4.0
yarl==1.8.1
